small changes and sterted get_cmdpath functions
